- Fix map not changing auto-zoom scale when entering builds at human settlements or changing tools - Progress on filtering out status.json writes when multi-boxing, but it does not work fully yet.

+++ SrvSurvey/game/JournalEntry.cs
@@ -58,7 +58,7 @@ namespace SrvSurvey
         public double FuelCapacity { get; set; }
         public bool StartLanded { get; set; }
         public string GameMode { get; set; }
-        public double Credits { get; set; }
+        public long Credits { get; set; }
         public double Loan { get; set; }
         public string language { get; set; }
         public string gameversion { get; set; }
